apps/ai-assistant/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.119.1
uvicorn==0.30.6
pydantic==2.8.2
requests==2.32.3
numpy<2.0.0
torch==2.3.1+cpu
sentence-transformers==2.6.1
faiss-cpu==1.7.4

